+++ common/src/main/java/io/github/mortuusars/exposure/integration/jei/ExposureJeiPlugin.java
@@ -67,7 +67,8 @@ public class ExposureJeiPlugin implements IModPlugin {
         ));
 
         if (Config.Client.SHOW_JEI_INFORMATION.get()) {
-            registration.addItemStackInfo(new ItemStack(Exposure.Items.PHOTOGRAPH_FRAME.get()),
+            registration.addItemStackInfo(List.of(new ItemStack(Exposure.Items.PHOTOGRAPH_FRAME.get()),
+                            new ItemStack(Exposure.Items.CLEAR_PHOTOGRAPH_FRAME.get())),
                     Component.translatable("exposure.jei.info.photograph_frame"));
         }